Kubernetes client object (#115)
* Build a Kubernetes client object based off of the local config

The client object is created from $HOME/.kube/config.

* Cluster valdation check using the Kubernetes client object

Now that we have the Kubernetes client object, use it to
connect to a running Kubernetes cluster to validate that
connection.

* Add further logging messages to client creation

* Create a client using the incluster config

Every pod is given the credentials to create a cluster
config. Use those credentials to contact Kubernetes
when inside a pod and fall back to local creds if
the file isn't present.