Replace get_images_from_org (#121)
* Get images from Dockerhub doing an http request

Currently we use a bash script to gather image date.
Let's pull that code into the broker so we have more
control over it.  This will help implement
async image gathering down the road.

* Remove get_image_for_org from the Dockerfile

This will help consolidate the image gathering
logic into one place, the broker.

* Remove the --scripts param and corresponding code

We no longer need to read anything from the scripts
directory so remove the --scripts parameter and the
code that used it.