Breakup all the Broker Clients into a clients pkg (#222)
* Break apart the Kubernetes Client

Make a seperate directory for clients so they are
not apart of the APB code.  This will make the clients
easy to initialize and create new ones.

* Break apart the Docker Client

* Break apart the Etcd client

The Etcd Client is a little tricker because it requires
a config file. Later when we unify all the clients we will
use individual functions just like etcd.

* Create a global client variable

Pull together all the client split work by using a global variable
to pass around the clients.