399 - APB Sandbox Role should be configurable (#403)
* APB Sandbox Role should be configurable

As mentioned in #399, there are apb's no and in the future that will
wish to create and/or modify service accounts. Until such a time as
apb's asking for the permissions (roles) that it will need, this moves
closer to that goal by making it a configurable item in the cluster
config section of the broker.

- Update example's and script's that generate a broker config to include
  a sandbox role (default: edit)
- Update apb package to pass roll on creation of sandbox

Fixes #399

* Test and document new configuration option

The broker's configuration documentation should mention the supported
options for the newly created sandbox role.

* Broker template should have sandbox role parameter

- Update the config docs for proper formatting and grammar
- Update deploy script to have sandbox_role option and the
  service-broker-template to have the sandbox_role parameter

* Remove role validation

It was pointed out by @jwmatthews that we shouldn't limit the acceptable
sandbox roles available as this is a workaround to make it so people who
need more than edit role can do so, but they are on their own to setup
the broker's role and sandbox role to get anything other than edit to
work.