Fix statements with formatting (#61)
If the formatting was in the middle of the string, the substition was
not done properly resulting in weird logs like the following:

  time="2018-04-23T17:54:45Z" level=info msg="Checking if namespace %s exists.blog-project"
  time="2018-04-23T17:54:45Z" level=info msg="Successfully created apb sandbox: [ %s ], with %s permissions in namespace %sapb-cdbfed9c-9c33-4a90-be5d-1582d9c338b4editdh-postgresql-apb-upda-jpvk4"

Using the *f version of the logger functions fixes the above. There were
some scenarios where the logs looked okay, but I believe that is because
the substition string was at the end of the log string. For example,
these statements look fine, but they also had the %s at the end of the
log string. I replaced these with the *f versions as well.

   time="2018-04-23T17:54:45Z" level=info msg="Spec.ID: 1dda1477cace09730bd8ed7a6505607e"
   time="2018-04-23T17:54:45Z" level=info msg="Spec.Name: dh-postgresql-apb"
   time="2018-04-23T17:54:45Z" level=info msg="Spec.Image: docker.io/jmrodri/postgresql-apb:demo"
   time="2018-04-23T17:54:45Z" level=info msg="Spec.Description: SCL PostgreSQL apb implementation"

I used this command to find the broken statements:

   git grep -n "log\." | grep "%" | grep -v "f("