Ensure we pass a path to client.NewRequest (#167)
We had an inconsistency where NewRequest was used in the apiv2 adapter,
we were passing in a full URL including the hostname instead of just the
path (which is what NewRequest expected). This caused the scheme://host
to be prefixed twice. Apparently http.client.Do was able to sanitize
this in some way with an implicit Parse call, but the problem came to
light when the URL was passed to a proxy to execute, and it wasn't
cleaned.